Use config to determine pydruid availability
